requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Pillow>=10.0.0
Werkzeug==2.3.7
pandas>=1.5.0
openpyxl>=3.0.0